refactor: reorganize inventory items into dedicated packages and improve item abstraction
- Moved inventory items to categorized packages (e.g., `navigator`, `pushback`, `visibility`, etc.) for better modularity.
- Converted `InventoryItem` to an abstract class to provide a default implementation for common properties and methods.
- Added new items: `ParkourItem`, `ProfileItem`, and `RewardsItem`, with respective configurations and functionality.
