+++ src/main/kotlin/dev/slne/surf/lobby/inventory/item/InventoryItem.kt
@@ -1,20 +1,21 @@
 package dev.slne.surf.lobby.inventory.item
 
-import dev.slne.surf.lobby.inventory.item.impl.NavigatorInventoryItem
-import dev.slne.surf.lobby.inventory.item.impl.PushbackDisableInventoryItem
-import dev.slne.surf.lobby.inventory.item.impl.PushbackEnableInventoryItem
-import dev.slne.surf.lobby.inventory.item.impl.ShowAllPlayersInventoryItem
-import dev.slne.surf.lobby.inventory.item.impl.ShowNonePlayersInventoryItem
-import dev.slne.surf.lobby.inventory.item.impl.ShowTeamPlayersInventoryItem
+import dev.slne.surf.lobby.inventory.item.impl.navigator.NavigatorInventoryItem
+import dev.slne.surf.lobby.inventory.item.impl.pushback.PushbackDisableInventoryItem
+import dev.slne.surf.lobby.inventory.item.impl.pushback.PushbackEnableInventoryItem
+import dev.slne.surf.lobby.inventory.item.impl.visibility.ShowAllPlayersInventoryItem
+import dev.slne.surf.lobby.inventory.item.impl.visibility.ShowNonePlayersInventoryItem
+import dev.slne.surf.lobby.inventory.item.impl.visibility.ShowTeamPlayersInventoryItem
 import org.bukkit.entity.Player
 import org.bukkit.inventory.ItemStack
 
-interface InventoryItem {
-    val slot: Int
+abstract class InventoryItem(
+    val slot: Int,
     val item: ItemStack
-    val permission: String?
+) {
+    val permission: String? = null
 
-    fun onInteract(player: Player)
+    fun onInteract(player: Player) {}
 
     companion object {
         val items = mutableListOf<InventoryItem>()
