+++ src/main/kotlin/dev/slne/surf/lobby/inventory/item/impl/parkour/ParkourItem.kt
@@ -1,0 +1,8 @@
+package dev.slne.surf.lobby.inventory.item.impl.parkour
+
+import dev.slne.surf.lobby.inventory.item.InventoryItem
+import org.bukkit.inventory.ItemType
+
+object ParkourItem : InventoryItem(0, ItemType.LEATHER_BOOTS.createItemStack().apply {
+
+})
