+++ src/main/kotlin/dev/slne/surf/lobby/inventory/item/impl/rewards/RewardsItem.kt
@@ -1,0 +1,8 @@
+package dev.slne.surf.lobby.inventory.item.impl.rewards
+
+import dev.slne.surf.lobby.inventory.item.InventoryItem
+import org.bukkit.inventory.ItemType
+
+object RewardsItem : InventoryItem(0, ItemType.GOLD_NUGGET.createItemStack().apply {
+
+})
