+++ src/main/kotlin/dev/slne/surf/lobby/inventory/item/impl/visibility/ShowNonePlayersInventoryItem.kt
@@ -1,4 +1,4 @@
-package dev.slne.surf.lobby.inventory.item.impl
+package dev.slne.surf.lobby.inventory.item.impl.visibility
 
 import dev.slne.surf.lobby.inventory.item.InventoryItem
 import dev.slne.surf.lobby.manager.PlayerVisibilityManager
@@ -9,10 +9,10 @@ import dev.slne.surf.surfapi.core.api.messages.adventure.sendText
 import org.bukkit.entity.Player
 import org.bukkit.inventory.ItemType
 
-object ShowTeamPlayersInventoryItem : InventoryItem {
+object ShowNonePlayersInventoryItem : InventoryItem {
     override val slot = 2
     override val permission: String = PermissionRegistry.PLAYER_VISIBILITY_ITEM
-    override val item = ItemType.ORANGE_DYE.createItemStack().apply {
+    override val item = ItemType.GRAY_DYE.createItemStack().apply {
         displayName {
             variableValue("Spieler-Sichtbarkeit")
         }
@@ -24,7 +24,7 @@ object ShowTeamPlayersInventoryItem : InventoryItem {
             }
             emptyLine()
             line {
-                variableValue("Nur Teammitglieder werden angezeigt")
+                error("Keine Spieler werden angezeigt")
             }
         }
     }
@@ -32,14 +32,14 @@ object ShowTeamPlayersInventoryItem : InventoryItem {
     override fun onInteract(player: Player) {
         PlayerVisibilityManager.setState(
             player.uniqueId,
-            PlayerVisibilityManager.VisibilityState.SHOW_NONE
+            PlayerVisibilityManager.VisibilityState.SHOW_ALL
         )
-        player.inventory.setItem(slot, ShowNonePlayersInventoryItem.item)
+        player.inventory.setItem(slot, ShowAllPlayersInventoryItem.item)
 
         player.sendText {
             appendPrefix()
             info("Du siehst jetzt ")
-            error("keine Spieler.")
+            success("alle Spieler.")
         }
     }
 }
